grade_programacao/requirements.txt
Django==1.11.16
django-tastypie==0.14.2
python-dateutil==2.7.4
python-mimeparse==1.6.0
pytz==2018.6
six==1.11.0
